==> cdrp/01-create-profiles-cellprofiler.ipynb <==
import pandas as pd
import os
import sklearn.metrics
import sys

sys.path.append("../profiling/")
import profiling

PROJECT_ROOT = "/dgx1nas1/cellpainting-datasets/CDRP/"
OUTPUT_FILE = "well_level_data_cellprofiler_1e-2.csv"
REG_PARAM = 1e-2
MATRIX_FILE = 'cos_cellprofiler_1e-2.csv'

# Load metadata
meta = pd.read_csv(os.path.join(PROJECT_ROOT, "inputs/metadata/index_trimmed_maxconc.csv"))
meta["broad_sample"] = meta["Treatment"].str.split("@", expand=True)[0]

prof24278 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24278))
prof24280 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24280))
prof24294 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24294))
prof24296 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24296))
prof24300 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24300))
prof24302 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24302))
prof24304 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24304))
prof24306 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24306))
prof24308 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24308))
prof24310 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24310))
prof24312 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24312))
prof24319 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24319))
prof24321 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24321))
prof24357 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24357))
prof25938 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25938))
prof25943 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25943))
prof25945 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25945))
prof25955 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25955))
prof25965 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25965))
prof25967 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25967))
prof25983 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25983))
prof25985 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25985))
prof25987 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25987))
prof25989 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25989))
prof25991 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25991))
prof26224 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(26224))
prof26239 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(26239))
prof24277 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24277))
prof24279 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24279))
prof24293 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24293))
prof24295 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24295))
prof24297 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24297))
prof24301 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24301))
prof24303 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24303))
prof24305 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24305))
prof24307 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24307))
prof24309 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24309))
prof24311 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24311))
prof24313 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24313))
prof24320 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24320))
prof24352 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(24352))
prof25937 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25937))
prof25939 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25939))
prof25944 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25944))
prof25949 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25949))
prof25962 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25962))
prof25966 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25966))
prof25968 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25968))
prof25984 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25984))
prof25986 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25986))
prof25988 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25988))
prof25990 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25990))
prof25992 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(25992))
prof26232 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(26232))
prof26247 = pd.read_csv("/dgx1nas1/cellpainting-datasets/CDRP/profiles_cp/{0}/{0}_normalized.csv".format(26247))

well_level_data = pd.concat([prof24278,prof24280,prof24294,prof24296,prof24300,prof24302,prof24304,prof24306,prof24308,prof24310,prof24312,prof24319,prof24321,prof24357,prof25938,prof25943,prof25945,prof25955,prof25965,prof25967,prof25983,prof25985,prof25987,prof25989,prof25991,prof26224,prof26239,prof24277,prof24279,prof24293,prof24295,prof24297,prof24301,prof24303,prof24305,prof24307,prof24309,prof24311,prof24313,prof24320,prof24352,prof25937,prof25939,prof25944,prof25949,prof25962,prof25966,prof25968,prof25984,prof25986,prof25988,prof25990,prof25992,prof26232,prof26247])
well_level_data['Treatment'] = well_level_data['Metadata_broad_sample'] + '@' + well_level_data['Metadata_mmoles_per_liter'].astype(str)
well_level_data.drop(['Metadata_Assay_Plate_Barcode', 'Metadata_Plate_Map_Name', 'Metadata_well_position', 'Metadata_mmoles_per_liter', 'Metadata_ASSAY_WELL_ROLE', 'Metadata_solvent', 'Metadata_pert_id', 'Metadata_pert_mfc_id', 'Metadata_pert_well',
                     'Metadata_pert_id_vendor', 'Metadata_cell_id', 'Metadata_broad_sample_type', 'Metadata_pert_vehicle', 'Metadata_pert_type' ], axis=1, inplace=True)

columns1 = ["Metadata_Plate", "Metadata_Well", "Metadata_broad_sample", "Treatment"]
columns2 = [i for i in well_level_data.columns if i not in columns1]
part1 = well_level_data.loc[:, columns1]
part2 = well_level_data.loc[:, columns2]

part2.dropna(axis='columns', inplace=True)

well_level_data = pd.concat([part1, part2],axis=1)
well_level_data.reset_index(inplace=True, drop=True)
columns2 = [i for i in well_level_data.columns if i not in columns1]

meta1 = meta.loc[:, ['Metadata_Plate', 'Metadata_Well', 'broad_sample_Replicate', 'Treatment']]
well_level_data = well_level_data.merge(meta1, how='right', left_on=['Metadata_Plate', 'Metadata_Well'], right_on=['Metadata_Plate', 'Metadata_Well']).drop_duplicates()
well_level_data.drop(columns=['Treatment_x'], inplace = True)
well_level_data.rename(columns = {'Treatment_y':'Treatment'}, inplace = True)
well_level_data = well_level_data.astype({'Metadata_Plate':'object', 'Metadata_Well':'object', 'Metadata_broad_sample':'object', 'broad_sample_Replicate':'object', 'Treatment':'object'})
well_level_data.reset_index(drop=True, inplace=True)

# 4. Sphering

whN = profiling.WhiteningNormalizer(well_level_data.loc[well_level_data["Metadata_broad_sample"].isin(["DMSO"]), columns2], REG_PARAM)

whD = whN.normalize(well_level_data[columns2])

# Save whitened profiles
well_level_data[columns2] = whD
well_level_data.to_csv(OUTPUT_FILE, index=False)

# 5. Treatment-level profiles / Mean Aggreagation

# Aggregate profiles
well_level_data["Metadata_broad_sample"] = well_level_data["Treatment"].str.split("@", expand=True)[0]
columns1 = ["Metadata_Plate", "Metadata_Well", "Treatment", "broad_sample_Replicate", "Metadata_broad_sample"]
columns2 = list(well_level_data.columns)[3:-2]
profiles = well_level_data.groupby("Treatment").mean().reset_index()

# Recover broad_sample column (cannot be used in groupby because it contains NaN values)
tmp = well_level_data.groupby(["Treatment", "Metadata_broad_sample"])["broad_sample_Replicate"].count().reset_index()
profiles = pd.merge(profiles.reset_index(drop=True), tmp, on="Treatment", how="left")

profiles = profiles[["Treatment", "Metadata_broad_sample"] + columns2]

# Remove samples without MOA (according to [1])
Y = pd.read_csv("data/CDRP_MOA_MATCHES_fixed_trimmed_filtered_v2.csv")
profiles = pd.merge(profiles, Y, left_on="Metadata_broad_sample", right_on="Var1")
profiles = profiles[["Treatment", "Metadata_broad_sample", "Metadata_moa.x"] + columns2].sort_values(by="Metadata_broad_sample")

# 6. Correlation matrix

# Compute Pearson correlation
COS = sklearn.metrics.pairwise.cosine_similarity(profiles[columns2], profiles[columns2])

# Transform to tidy format
df = pd.DataFrame(data=COS, index=list(profiles.Metadata_broad_sample), columns=list(profiles.Metadata_broad_sample))
df = df.reset_index().melt(id_vars=["index"])

# Annotate rows
df2 = pd.merge(
    df, 
    profiles[["Metadata_broad_sample", "Metadata_moa.x"]], 
    how="left", 
    left_on="index", # <=== Rows
    right_on="Metadata_broad_sample"
).drop("Metadata_broad_sample",axis=1)

# Annotate columns
df2 = pd.merge(
    df2, profiles[["Metadata_broad_sample", "Metadata_moa.x"]],
    how="left", 
    left_on="variable", # <=== Columns
    right_on="Metadata_broad_sample"
).drop("Metadata_broad_sample",axis=1)

# Rename columns and save
df2.columns = ["Var1", "Var2", "value", "Metadata_moa.x", "Metadata_moa.y"]
df2.to_csv(MATRIX_FILE)
