==> cdrp/02-profiles-visualizations.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sb
import umap
import sys

sys.path.append("../profiling/")
import profiling

NUM_FEATURES = 6400
INPUT_PROFILES = "data/well_level_data.csv"

wells = pd.read_csv(INPUT_PROFILES)
columns2 = [str(i) for i in range(NUM_FEATURES)]

# Well-level UMAP

reducer = umap.UMAP()
embeddings = reducer.fit_transform(wells[columns2])
aux = pd.concat((pd.DataFrame(embeddings, columns=["X", "Y"]), wells.reset_index()), axis=1)

aux["Control"] = aux["Treatment"] == "NA@NA" #["broad_sample"].isna()
plt.figure(figsize=(10,10))
sb.scatterplot(data=aux, x="X", y="Y", hue="Control")

# Treatment-level correlations

X, Y = profiling.load_correlation_matrix("data/correlation_matrix.csv")

sb.clustermap(X[Y.Var1])

# MOA counts

Y.groupby("Metadata_moa.x")["Var1"].count()
